Update getUserNumberInput function to fix input retrieval issue
Modified the getUserNumberInput function in app.js to resolve the error caused by not being able to retrieve the input value. Replaced 'usrInput' with 'input-number' as the correct ID for the input element in the HTML file. Now the function retrieves the user input correctly, allowing the calculator functions to work properly.